move matching logic over to new class URLMatcher
